requirements.txt
google-genai
python-dotenv
feedparser
Pillow
requests
